+++ monticore-generator/src/main/java/de/monticore/codegen/symboltable/CommonScopeGenerator.java
@@ -40,10 +40,6 @@ public class CommonScopeGenerator implements ScopeGenerator {
         getSimpleName(scopeName + GeneratorHelper.BUILDER),
         genHelper.getTargetPackage(), handCodedPath);
 
-    String serializerName = getSimpleTypeNameToGenerate(
-        getSimpleName(getSimpleName(scopeName) + GeneratorHelper.SERIALIZER),
-        genHelper.getTargetPackage(), handCodedPath);
-
     String interfaceName = "I" + className;
 
     // Maps Symbol Name to Symbol Kind Name
@@ -87,7 +83,7 @@ public class CommonScopeGenerator implements ScopeGenerator {
       }
     }
     
-    // list of superscopevisitorss that the scope must accept
+    // list of superscopevisitors that the scope must accept
     Set<String> superScopeVisitors = new HashSet<>();
     for (CDSymbol cdSymbol : genHelper.getAllSuperCds(genHelper.getCd())) {
       String qualifiedScopeVisitorName = genHelper.getQualifiedScopeVisitorType(cdSymbol);
@@ -100,8 +96,6 @@ public class CommonScopeGenerator implements ScopeGenerator {
         className + ".java");
     final Path builderFilePath = Paths.get(Names.getPathFromPackage(genHelper.getTargetPackage()),
         builderName + ".java");
-    final Path serializerFilePath = Paths
-        .get(Names.getPathFromPackage(genHelper.getTargetPackage()), serializerName + ".java");
     final Path interfaceFilePath = Paths
         .get(Names.getPathFromPackage(genHelper.getTargetPackage()), interfaceName + ".java");
 
@@ -110,8 +104,6 @@ public class CommonScopeGenerator implements ScopeGenerator {
     genEngine.generateNoA("symboltable.Scope", filePath, className, scopeRule, symbolNamesWithSuperGrammar, superScopeVisitors);
     genEngine.generateNoA("symboltable.ScopeInterface", interfaceFilePath, interfaceName, symbolNames, superScopes);
     genEngine.generateNoA("symboltable.ScopeBuilder", builderFilePath, builderName,
-        scopeName + GeneratorHelper.BUILDER);
-    genEngine.generateNoA("symboltable.serialization.ScopeSerialization", serializerFilePath,
-        serializerName, getSimpleName(scopeName));
+        scopeName);
   }
 }