+++ monticore-generator/src/main/resources/de/monticore/monticore_noemf.groovy
@@ -21,7 +21,7 @@ Log.debug("Handcoded files     : " + handcodedPath, LOG_ID)
 IncrementalChecker.initialize(out, report)
 InputOutputFilesReporter.resetModelToArtifactMap()
 globalScope = createGlobalScope(modelPath)
-Reporting.init(report.getAbsolutePath(), reportManagerFactory)
+Reporting.init(out.getAbsolutePath(), report.getAbsolutePath(), reportManagerFactory)
 // ############################################################
 
 // ############################################################
@@ -79,7 +79,7 @@ for (astGrammar in getParsedGrammars()) {
   generateSymbolTable(glex, astGrammar, globalScope, astClassDiagram, out, handcodedPath)
   
   // M9 Generate ast classes, visitor and context condition
-  generate(glex, globalScope, astClassDiagram, out, templatePath)
+  generate(glex, globalScope, astClassDiagram, out, templatePath, handcodedPath)
 
   Log.info("Grammar " + astGrammar.getName() + " processed successfully!", LOG_ID)