+++ monticore-grammar/src/main/java/de/monticore/expressions/prettyprint2/JavaClassExpressionsPrettyPrinter.java
@@ -8,13 +8,14 @@ import de.monticore.expressions.javaclassexpressions._visitor.JavaClassExpressio
 import de.monticore.prettyprint.CommentPrettyPrinter;
 import de.monticore.prettyprint.IndentPrinter;
 
-public class JavaClassExpressionsPrettyPrinter implements JavaClassExpressionsVisitor {
+public class JavaClassExpressionsPrettyPrinter extends CommonExpressionsPrettyPrinter implements JavaClassExpressionsVisitor {
   
  protected JavaClassExpressionsVisitor realThis;
   
   protected IndentPrinter printer;
   
   public JavaClassExpressionsPrettyPrinter(IndentPrinter printer) {
+    super(printer);
     this.printer=printer;
     realThis = this;
 
@@ -26,7 +27,7 @@ public class JavaClassExpressionsPrettyPrinter implements JavaClassExpressionsVi
     getPrinter().print("super");
     CommentPrettyPrinter.printPostComments(node, getPrinter());
   }
-  
+
   @Override
   public void handle(ASTSuperSuffix node) {
     CommentPrettyPrinter.printPreComments(node, getPrinter());
@@ -45,7 +46,7 @@ public class JavaClassExpressionsPrettyPrinter implements JavaClassExpressionsVi
     }
     CommentPrettyPrinter.printPostComments(node, getPrinter());
   }
-  
+
   @Override
   public void handle(ASTSuperExpression node) {
     CommentPrettyPrinter.printPreComments(node, getPrinter());
@@ -102,14 +103,15 @@ public class JavaClassExpressionsPrettyPrinter implements JavaClassExpressionsVi
     handle(node.getPrimaryGenericInvocationExpression());
     CommentPrettyPrinter.printPostComments(node, getPrinter());
   }
-  
+
   @Override
-  public void handle(ASTNameExpression node) {
+  public void handle(ASTPrimaryGenericInvocationExpression node) {
     CommentPrettyPrinter.printPreComments(node, getPrinter());
-   getPrinter().print(node.getName());
-   CommentPrettyPrinter.printPostComments(node, getPrinter());
+    node.getExtTypeArguments().accept(getRealThis());
+    node.getGenericInvocationSuffix().accept(getRealThis());
+    CommentPrettyPrinter.printPostComments(node, getPrinter());
   }
-  
+
   @Override
   public void handle(ASTInstanceofExpression node) {
     CommentPrettyPrinter.printPreComments(node, getPrinter());
@@ -118,25 +120,34 @@ public class JavaClassExpressionsPrettyPrinter implements JavaClassExpressionsVi
     node.getExtType().accept(getRealThis());
     CommentPrettyPrinter.printPostComments(node, getPrinter());
   }
-  
+
   @Override
-  public void handle(ASTArguments node) {
+  public void handle(ASTThisExpression node) {
     CommentPrettyPrinter.printPreComments(node, getPrinter());
-    getPrinter().print("(");
-    int count = 0;
-    if (!node.isEmptyExpressions()) {
-      for (ASTExpression ast : node.getExpressionList()) {
-        if (count > 0) {
-          getPrinter().print(",");
-        }
-        ast.accept(getRealThis());
-        count++;
-      }
-    }
-    getPrinter().print(")");
+    node.getExpression().accept(getRealThis());
+    getPrinter().print(".");
+    getPrinter().print("this");
     CommentPrettyPrinter.printPostComments(node, getPrinter());
   }
-  
+
+  @Override
+  public void handle(ASTArrayExpression node) {
+    CommentPrettyPrinter.printPreComments(node, getPrinter());
+    node.getExpression().accept(getRealThis());
+    getPrinter().print("[");
+    node.getIndexExpression().accept(getRealThis());
+    ;
+    getPrinter().print("]");
+    CommentPrettyPrinter.printPostComments(node, getPrinter());
+  }
+
+  @Override
+  public void visit(ASTPrimaryThisExpression node) {
+    CommentPrettyPrinter.printPreComments(node, getPrinter());
+    getPrinter().print("this");
+    CommentPrettyPrinter.printPostComments(node, getPrinter());
+  }
+
   public IndentPrinter getPrinter() {
     return this.printer;
   }