+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Release Notes
 
-##  MontiCore 5.0.4
+##  MontiCore 5.0.5
+- The IncGen-reports are stored in the source code directory
+- Removed MutableScope
+- IncGen-Reports are stored 
 - Removed deprecated keyword _ast_ (use _astrule_) in *.mc4
+- Add visitors for symbol table
+- Enable TOP mechanism for visitors.
 
 ##  MontiCore 5.0.3
 - Use the following emf coordinates (MB):